refactor(chat): add localChat selectors
